fix: correct dtype, chat template, transformers version, and stale docs
- torch_dtype float32 → bfloat16 (halves base model VRAM: ~4.5GB → ~2.3GB)
- transformers>=4.45.0 → >=4.46.0 (LFM2 added July 2025, PR #39340)
- AlpacaDataset: raw alpaca format → apply_chat_template with response-only
  label masking; also handles optional 'input' field in alpaca records
- PLAN.md: LFM2-1.2B → LFM2.5-1.2B-Instruct, hidden_size 2560 → 2048 throughout

+++ experiments/baseline_finetuning.py
@@ -66,7 +66,7 @@ def main() -> None:
 
     print("Loading LFM2.5-1.2B-Instruct for naive fine-tuning...")
     model = AutoModelForCausalLM.from_pretrained(
-        "LiquidAI/LFM2.5-1.2B-Instruct", torch_dtype=torch.float32
+        "LiquidAI/LFM2.5-1.2B-Instruct", torch_dtype=torch.bfloat16
     ).to(device)
     tok = AutoTokenizer.from_pretrained("LiquidAI/LFM2.5-1.2B-Instruct")
     if tok.pad_token is None:
